troubleshooting: fix details summary layout
This PR fixed the details summary layout because the kramdown's
parser regard the summary as a block and generates unpexpected
HTML layout like the following.

So we explicitly specified the summary is a inline block to solve
it. After https://github.com/gettalong/kramdown/pull/819 is solved,
we can just add the parse-block-html option to handdle it.