troubleshooting: fix details summary layout (#399)
This PR fixed the details summary layout because the kramdown's
parser regards the summary as a block and generates unexpected
HTML layout like the following.

So we explicitly specified the summary is a inline block to solve
it. After https://github.com/gettalong/kramdown/pull/819 is solved,
we can just add the `parse-block-html` option to handle it.

|Before|After|
|---|---|
| ![Screenshot from 2024-12-12
09-09-21](https://github.com/user-attachments/assets/771496ad-1de4-4eb0-8d05-3d224f5a9f58)
| ![Screenshot from 2024-12-12
09-10-01](https://github.com/user-attachments/assets/3d97b8d6-5fb9-4567-b223-6c18eb070a83)
|

FYI: This PR doesn't add additional css to fix the layout of details and
summary.